// app/types/favorites.ts
export interface ProductForFavorites {
  id: number;
  titulo: string;
  slug: string;
  imageUrl: string;
  // Aquí podrías añadir más campos si fueran necesarios para el ProductCard
  // por ejemplo, una breve descripción o el objeto "imagen" completo
  // si ProductCard lo necesitara. Por ahora, nos mantenemos con lo esencial
  // para el enlace y la visualización básica.
}